requirements.txt
# Telegram Bot Framework
aiogram>=2.25.1

# Database
SQLAlchemy>=2.0.0
alembic>=1.12.0
psycopg2-binary>=2.9.9  # PostgreSQL adapter

# OCR & PDF Processing
pytesseract>=0.3.10
Pillow>=10.0.0  # For image processing
pdf2image>=1.16.3  # For converting PDF to images
pdfplumber>=0.10.2  # For PDF text extraction

# Utilities
python-dotenv>=1.0.0  # For environment variables
pytz>=2023.3  # For timezone handling
aiofiles>=23.2.1  # For async file operations
pydantic>=2.0.0  # For data validation

# Logging
structlog>=23.1.0

# Development tools
black>=23.9.1  # Code formatter
flake8>=6.1.0  # Code linter
pytest>=7.4.2  # Testing
pytest-asyncio>=0.21.1  # Async testing